resolves #231
