+++ iOS-Email-Client/Libs/Utils.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-class Utils{
+class Utils: SharedUtils {
     
     class func createLabelWithDynamicHeight(_ width: CGFloat, _ fontSize: CGFloat) -> UILabel {
         let label:UILabel = UILabel(frame: CGRect(x: 0, y: 0, width: width, height: CGFloat.greatestFiniteMagnitude))
@@ -38,26 +38,6 @@ class Utils{
         label.text = text
         label.sizeToFit()
         return label.frame.height/label.font.pointSize
-    }
-    
-    class func generateRandomColor() -> UIColor {
-        let hue : CGFloat = CGFloat(arc4random() % 256) / 256
-        let saturation : CGFloat = CGFloat(arc4random() % 128) / 256 + 0.5
-        let brightness : CGFloat = CGFloat(arc4random() % 128) / 256 + 0.5
-        
-        return UIColor(hue: hue, saturation: saturation, brightness: brightness, alpha: 1)
-    }
-    
-    class func convertToDictionary(text: String) -> [String: Any]? {
-        guard let data = text.data(using: .utf8) else {
-            return nil
-        }
-        do {
-            return try JSONSerialization.jsonObject(with: data, options: []) as? [String: Any]
-        } catch {
-            print(error.localizedDescription)
-        }
-        return nil
     }
 
     class func convertToJSONString(dictionary: [String: Any]) -> String? {
