+++ iOS-Email-Client/Shared/SharedUtils.swift
@@ -1,0 +1,32 @@
+//
+//  SharedUtils.swift
+//  iOS-Email-Client
+//
+//  Created by Allisson on 11/8/18.
+//  Copyright © 2018 Criptext Inc. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class SharedUtils {
+    class func generateRandomColor() -> UIColor {
+        let hue : CGFloat = CGFloat(arc4random() % 256) / 256
+        let saturation : CGFloat = CGFloat(arc4random() % 128) / 256 + 0.5
+        let brightness : CGFloat = CGFloat(arc4random() % 128) / 256 + 0.5
+        
+        return UIColor(hue: hue, saturation: saturation, brightness: brightness, alpha: 1)
+    }
+    
+    class func convertToDictionary(text: String) -> [String: Any]? {
+        guard let data = text.data(using: .utf8) else {
+            return nil
+        }
+        do {
+            return try JSONSerialization.jsonObject(with: data, options: []) as? [String: Any]
+        } catch {
+            print(error.localizedDescription)
+        }
+        return nil
+    }
+}
